extract wrapBigNumber from utils/common
